First working draft of CEM transmission/reception
This commit introduces a first working draft of the CEM tx/rx.
Some details still need to be defined (e.g. the factors added
inside ceBasicService.h) and it was necessary to manually
modify some asn1c-generated files to enable the communication
(this will not be needed after the factors and the ranges are
well defined and the final version of the .asn file is ready).