Use SUMO in v2v-80211p-gps-raw-tc-example.cc
With this commit, the v2v-80211p-gps-raw-tc-example example now
relies of SUMO to manage the mobility of the nodes, instead of
using GPS-tc.
This choice comes from the fact that using SUMO provides more
flexibility for the initial testing and evaluation of the CE
Basic Service and of the transmission and reception of CEMs,
even if the positions of vehicles are actually simulated and
not coming from real GPS traces.
Using SUMO works as long as we are interested in the network
parameters and not on the actual GNSS raw data exchanged between
vehicles, which would require synchronized GNSS traces and
raw data traces (including a different raw data trace for each
vehicle, which is not available yet - now a single realistic
raw trace is used for all the vehicles with some initial CEM
dissemination delay to simulate the real case in which all the
vehicles will send the I and D frames at slightly different times).